+++ cachalot/__init__.py
@@ -1,10 +1,4 @@
-VERSION = (2, 5, 3)
+VERSION = (2, 6, 0)
 __version__ = ".".join(map(str, VERSION))
 
-try:
-    from django import VERSION as DJANGO_VERSION
-
-    if DJANGO_VERSION < (3, 2):
-        default_app_config = "cachalot.apps.CachalotConfig"
-except ImportError:  # pragma: no cover
-    default_app_config = "cachalot.apps.CachalotConfig"
+default_app_config = "cachalot.apps.CachalotConfig"