+++ cachalot/tests/test_utils.py
@@ -1,6 +1,8 @@
-from django import VERSION as DJANGO_VERSION
 from django.core.management.color import no_style
-from django.db import connection, transaction
+from django.db import DEFAULT_DB_ALIAS, connection, connections, transaction
+
+from django.test import TransactionTestCase
+from django.test.utils import CaptureQueriesContext
 
 from ..utils import _get_tables
 from .models import PostgresModel
@@ -11,7 +13,6 @@ class TestUtilsMixin:
         self.is_sqlite = connection.vendor == 'sqlite'
         self.is_mysql = connection.vendor == 'mysql'
         self.is_postgresql = connection.vendor == 'postgresql'
-        self.django_version = DJANGO_VERSION
         self.force_reopen_connection()
 
     # TODO: Remove this workaround when this issue is fixed:
@@ -19,8 +20,6 @@ class TestUtilsMixin:
     def tearDown(self):
         if connection.vendor == 'postgresql':
             flush_args = [no_style(), (PostgresModel._meta.db_table,),]
-            if float(".".join(map(str, DJANGO_VERSION[:2]))) < 3.1:
-                flush_args.append(())
             flush_sql_list = connection.ops.sql_flush(*flush_args)
             with transaction.atomic():
                 for sql in flush_sql_list:
@@ -61,3 +60,61 @@ class TestUtilsMixin:
         assert_function(data2, data1)
         if result is not None:
             assert_function(data2, result)
+
+class FilteredTransactionTestCase(TransactionTestCase):
+    """
+    TransactionTestCase with assertNumQueries that ignores BEGIN, COMMIT and ROLLBACK
+    queries.
+    """
+    def assertNumQueries(self, num, func=None, *args, using=DEFAULT_DB_ALIAS, **kwargs):
+        conn = connections[using]
+
+        context = FilteredAssertNumQueriesContext(self, num, conn)
+        if func is None:
+            return context
+
+        with context:
+            func(*args, **kwargs)
+
+
+class FilteredAssertNumQueriesContext(CaptureQueriesContext):
+    """
+    Capture queries and assert their number ignoring BEGIN, COMMIT and ROLLBACK queries.
+    """
+    EXCLUDE = ('BEGIN', 'COMMIT', 'ROLLBACK')
+
+    def __init__(self, test_case, num, connection):
+        self.test_case = test_case
+        self.num = num
+        super().__init__(connection)
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        super().__exit__(exc_type, exc_value, traceback)
+        if exc_type is not None:
+            return
+
+        filtered_queries = []
+        excluded_queries = []
+        for q in self.captured_queries:
+            if q['sql'].upper() not in self.EXCLUDE:
+                filtered_queries.append(q)
+            else:
+                excluded_queries.append(q)
+
+        executed = len(filtered_queries)
+
+        self.test_case.assertEqual(
+            executed,
+            self.num,
+            f"\n{executed} queries executed on {self.connection.vendor}, {self.num} expected\n" +
+            "\nCaptured queries were:\n" +
+            "".join(
+                f"{i}. {query['sql']}\n"
+                for i, query in enumerate(filtered_queries, start=1)
+            ) +
+            "\nCaptured queries, that were excluded:\n" +
+            "".join(
+                f"{i}. {query['sql']}\n"
+                for i, query in enumerate(excluded_queries, start=1)
+            )
+        )