+++ cachalot/utils.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 from uuid import UUID
 
 from django.contrib.postgres.functions import TransactionNow
-from django.db import connections
+from django.db import connections, connection
 from django.db.models import Exists, QuerySet, Subquery
 from django.db.models.expressions import RawSQL
 from django.db.models.functions import Now
@@ -35,24 +35,42 @@ CACHABLE_PARAM_TYPES = {
 }
 UNCACHABLE_FUNCS = {Now, TransactionNow}
 
-try:
-    # TODO Drop after Dj30 drop
-    from django.contrib.postgres.fields.jsonb import JsonAdapter
-    CACHABLE_PARAM_TYPES.update((JsonAdapter,))
-except ImportError:
-    pass
+# Check if psycopg2 or 3 is used
+connection.ensure_connection()
+underlying_connection = connection.connection
 
-try:
+if underlying_connection.__class__.__module__.startswith('psycopg2'):
     from psycopg2 import Binary
     from psycopg2.extras import (
         NumericRange, DateRange, DateTimeRange, DateTimeTZRange, Inet, Json)
-except ImportError:
-    pass
-else:
     CACHABLE_PARAM_TYPES.update((
         Binary, NumericRange, DateRange, DateTimeRange, DateTimeTZRange, Inet,
         Json,))
+elif underlying_connection.__class__.__module__.startswith('psycopg'):
+    from django.db.backends.postgresql.psycopg_any import (
+        NumericRange, DateRange, DateTimeRange, DateTimeTZRange, Inet,
+    )
 
+    from psycopg.dbapi20 import Binary
+
+    from psycopg.types.numeric import (
+        Int2, Int4, Int8, Float4, Float8,
+    )
+
+    from ipaddress import (
+        IPv4Address,
+        IPv6Address,
+    )
+
+    from psycopg.types.json import (
+        Json, Jsonb,
+    )
+
+    CACHABLE_PARAM_TYPES.update((
+        NumericRange, DateRange, DateTimeRange, DateTimeTZRange, Inet, Json, Jsonb,
+        Int2, Int4, Int8, Float4, Float8, IPv4Address, IPv6Address,
+        Binary,
+    ))
 
 def check_parameter_types(params):
     for p in params:
@@ -131,13 +149,7 @@ def _find_rhs_lhs_subquery(side):
     elif h_class is QuerySet:
         return side.query
     elif h_class in (Subquery, Exists):  # Subquery allows QuerySet & Query
-        try:
-            return side.query.query if side.query.__class__ is QuerySet else side.query
-        except AttributeError:  # TODO Remove try/except closure after drop Django 2.2
-            try:
-                return side.queryset.query
-            except AttributeError:
-                return None
+        return side.query.query if side.query.__class__ is QuerySet else side.query
     elif h_class in UNCACHABLE_FUNCS:
         raise UncachableQuery