+++ CHANGELOG.rst
@@ -1,12 +1,23 @@
 What’s new in django-cachalot?
 ==============================
 
+2.6.0
+-----
+
+- Dropped Django 2.2 and 4.0 support
+- Added Django 4.2 and Python 3.11 support
+- Added psycopg support (#229)
+- Updated tests to account for the `BEGIN` and `COMMIT` query changes in Django 4.2
+- Standardized django version comparisons in tests
+
 2.5.3
 -----
+
 - Verify get_meta isn't none before requesting db_table (#225 #226)
 
 2.5.2
 -----
+
 - Added Django 4.1 support (#217)
 
 2.5.1